Ensure recovery from failure to start upload session tasks when there is no session
When there is no session (due to recently invalidated / recreated session, typically from transition between background / active states), fail the upload so that we'll stop reading and then start again when the session is ensured

NOTE: We did correctly handle the failure case when unable to create POST body for the metadata and samples, by throwing an error. However, in the case of no valid session, we simply logged the error and ate it. Now we correctly call the completion handler to indicate the error, and the upload manager handles this by stopping the read, and it will start again when session is recreated